requirements.txt
beautifulsoup4==4.11.1
blis==0.7.7
bs4==0.0.1
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
et-xmlfile==1.1.0
fonttools==4.32.0
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
murmurhash==1.0.6
nltk==3.5
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.0.5
pathy==0.6.1
Pillow==9.1.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redash-api-client==0.3.0
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2
spacy==3.2.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
thinc==8.0.15
threadpoolctl==3.1.0
tqdm==4.47.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
xlrd==1.2.0
